Moved api schemas to their appropriate location in base.utils.api rather than polluting the main namespace
