+++ oioioi/contests/api.py
@@ -8,8 +8,8 @@ from rest_framework.parsers import MultiPartParser
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from oioioi.api_schemas import contest_id_parameter
 from oioioi.base.permissions import enforce_condition, not_anonymous
+from oioioi.base.utils.api import contest_id_parameter
 from oioioi.contests.controllers import submission_template_context
 from oioioi.contests.forms import SubmissionFormForProblemInstance
 from oioioi.contests.models import Contest, Submission
